I think this might be everything?
- fully merge mtarp into arptypes since it only had types
- move gen-ts into arptypes
- move SerializedCreature into arptypes
