+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["foldertree", "indexed", "nonempty", "arpeggio", "worker", "mtarp", "arptypes"]
+members = ["foldertree", "indexed", "nonempty", "arpeggio", "worker", "arptypes"]
 exclude = ["arpui"]
 resolver = "2"
