+++ arptypes/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+anyhow = { workspace = true }
 derive_more = { workspace = true }
 foldertree = { path = "../foldertree", version = "*", features = ["serde"] }
 indexed = { version = "*", path = "../indexed", features = ["serde"] }
@@ -14,6 +15,7 @@ num = { workspace = true }
 num-traits = { workspace = true }
 serde = { workspace = true }
 serde_derive = { workspace = true }
+strum = { workspace = true }
 thiserror = { workspace = true }
 ts-rs = { version = "7.0.0", features = ["serde-compat", "uuid-impl"] }
 uom = { workspace = true }
