+++ arptypes/src/multitenant.rs
@@ -1,9 +1,9 @@
 use serde::{Deserialize, Serialize};
 use ts_rs::TS;
 use uuid::Uuid;
+use strum;
 
-use arpeggio::types::{AbilityID, CreatureID, GMCommand, PlayerCommand, PlayerID, Point3, SceneID};
-use arptypes::uuid_id;
+use crate::{uuid_id, AbilityID, CreatureID, GMCommand, PlayerCommand, PlayerID, Point3, SceneID};
 
 uuid_id!(GameID);
 uuid_id!(InvitationID);
