--- mtarp/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "mtarp"
-version = "0.1.0"
-authors = ["Christopher Armstrong"]
-edition = "2021"
-description = "Multi-Tenant Arpeggio -- the types & behavior for managing a service that runs multiple games simultaneously."
-
-[dependencies]
-anyhow = { workspace = true }
-arpeggio = { version = "*", path = "../arpeggio" }
-arptypes = { path = "../arptypes", version = "*" }
-async-trait = "0.1.73"
-clap = { version = "4.4.6", features = ["derive"] }
-foldertree = { version = "*", path = "../foldertree" }
-futures = "0.3.1"
-http = "0.2.0"
-serde = "1.0"
-serde_derive = "1.0"
-serde_json = "1.0"
-serde_yaml = "0.8"
-strum = { workspace = true }
-thiserror = { workspace = true }
-tracing = { workspace = true }
-tracing-subscriber = { version = "0.3.17", features = ["env-filter", "json"] }
-ts-rs = { version = "7.0.0", features = ["serde-compat", "uuid-impl"] }
-# Why do I have this uom dependency here?
-uom = { version = "0.35", default_features = false }
-uuid = { workspace = true }
