+++ worker/Cargo.toml
@@ -13,6 +13,7 @@ crate-type = ["cdylib"]
 [dependencies]
 anyhow = { workspace = true }
 arpeggio = { version = "*", path = "../arpeggio" }
+arptypes = { path = "../arptypes", version = "*" }
 console_error_panic_hook = "0.1.6"
 foldertree = { path = "../foldertree" }
 futures-util = { version = "0.3.28", default-features = false }
@@ -20,9 +21,8 @@ futures-util = { version = "0.3.28", default-features = false }
 getrandom = { version = "0.2.9", features = ["js"] }
 gloo-timers = "0.3.0"
 google-oauth = { version = "1.6.0", features = ["wasm"] }
-mtarp = { version = "*", path = "../mtarp" }
 percent-encoding = { workspace = true }
-reqwest = { version = "0.11.22", features = [ "json", "multipart" ]}
+reqwest = { version = "0.11.22", features = ["json", "multipart"] }
 serde = { workspace = true }
 serde-wasm-bindgen = "0.6.0"
 serde_json = "1.0"
