+++ worker/src/cfworker.rs
@@ -5,7 +5,7 @@ use tracing::{error, info};
 use worker::{event, Context, Cors, Env, Method, Request, Response, Result};
 
 use crate::{anyhow_str, rust_error, storage};
-use mtarp::types::{GameID, GameList, GameMetadata, GameProfile, Role, UserID};
+use arptypes::multitenant::{GameID, GameList, GameMetadata, GameProfile, Role, UserID};
 
 /// The main cloudflare Worker for Arpeggio. Handles routes for listing &
 /// creating games, etc, and forwarding websockets to the Durable Object.
